Cleans ups Vee module and operator
Removes a significant number of old/unused functions and refactors some of the Vee module. Some functions are combined. Outputs also clarified and inputs adjusted.
